#48 play sound when a message is received. put all chat code in one place.
